Move paragraph about WarnPlaintextPorts to the section about Tor.
From an end-user perspective, it surely makes sense to put that in the section
about Claws Mail, since that's where user-visible effects can be seen. However,
this is not exactly the audience our design doc is written: here, the idea is to
explain why Tor is configured in this or that way, so this paragraph belongs to
the section about Tor configuration.

Besides, the section about Tor configuration is the place where we link
to the actual bits this paragraph is about. If we wanted to leave it in the
Claws Mail section, then it would require adding a link to `torrc' there.