Revert disputable phrasing change on the press page.
This change was kinda hidden in 86a79d4 ("correct link [...]") and was meant to
make the phrasing more consistent with what we use elsewhere. More specifically,
this new phrasing was copied from some page aimed at sysadmins. I personally
think that it's less appropriate, for the target audience of this section, than
the previous one, hence this revert.