grub.cfg: don't search for "root".
This searching introduced a security issue: we would have happily booted any
suitably named Linux kernel stored on a GPT partition that contains
a /live/Tails.module file.

It's not necessary, as apparently GRUB2 initializes $root to something based on
the device where it was loaded from initially, which is exactly what we need.

So, let's get rid of it!