Convert windows_camouflage.feature to the new snapshot model.
So far this should be the only .feature that will run slower with the
new snapshot model, since we do not have the snapshot saved *after*
setting the Windows camouflage Tails Greeter option and logging in. We
could of course consider introducing a snapshot for that.

Crazy idea to consider: in the snapshot description structs, we could
introduce a "per-feature" flag that when true means that the snapshot
will be removed when the feature ends. That way we could have more
suitable "full background" snapshots.