Don't modify files in the source tree before having possibly merged the base branch into it.
Previously, when building in Jenkins, we refreshed the wiki (twice) and
translations before merging the base branch. As a result, we're seeing build
failures with the "Your local changes to the following files would be
overwritten by merge" error message.

Let's now build the wiki and refresh translations only _after_ having merged the
base branch, which makes more sense anyway: we want to include changes from the
base branch into the wiki and translations. And while we're at it, let's refresh
the wiki only once when building in Jenkins.

Note that Vagrant builds will still build the wiki before setting up the ramdisk
and running live-build: those builds don't merge the base branch anyway, so they
can as well keep the wiki build caching mechanism, which is a nice optimization.

Will-fix: #9406