Enable Spice in the guest.
This seems to help with the issues we have with lost mouse events, in
particular in combination with an absolute pointing device, like the
'tablet' we already have.

There are some interesting bits about this mentioned in the Spice
manual:

    https://elmarco.fedorapeople.org/manual.html

So currently we use mouse mode='client' which "is appropriate for
... a loaded server, since cursor has smooth motion and
responsiveness", which we have some data suggestion have been
affecting us (see discussion on #8928). However, it also says "the
cursor might lose synchronization (position and shape) for a while"
but it's not mentioned what may cause this (when there's high
load?). It could be interesting to test mouse mode='server' too, just
for comparison.

Will-fix: #9425