Always export TMPDIR back to the test suite's shell environment.
We should do this so that the default TMPDIR (currently
/tmp/TailsToaster, as set in features/config/defaults.yml) is exported
back to the environment. This way TMPDIR is always synchronized
internally in the test suite and in the shell environment; without
this change it's synchronized like that except in the default case,
when TMPDIR wasn't in the env already, and --tmpdir wasn't passed,
which is confusing, and leads to subprocesses (e.g. guestfs) using
/tmp as the temporary directory.

Will-fix: #9479