Pidgin AppArmor profile: disable launchpad-integration abstraction.
That abstraction gives /usr/bin/launchpad-integration wide open access to many
files and executables; on the one hand we don't care much, since we don't ship
that binary. On the other hand, some of those wide-open rules (for example,
those about /** and /{,usr/}lib*/{,**/}*.so{,.*}) won't play well with aliases:
they make the policy needlessly hard to audit, and may increase its
compilation time.

The Pidgin profile is the only one we currently ship that includes the
launchpad-integration abstraction. Same on my current Debian sid, so we should
be good at least for Tails/Jessie (and possibly even for Tails/Stretch).