AppArmor: don't allow Tor Browser to access machine-id.
It was initially added because it was needed for Tor Browser to play sound with
PulseAudio, but in the end it turns out that this use case works just fine
without allowing such access => let's tighten thing a bit, and make our delta
with upstream's profile smaller.

Note that machine-id is currently a per-Tails-version identifier, but depending
on the outcome of #7100 it may become a per-Tails-boot one.