Clarify the relationship between Tails and the original OS (#9435)
We don't have versions of Tails *for* Windows or Mac as Tails runs
independently from them. This is only the OS we are *installing from*.

If it's not listed there (eg. Android) then you can't install Tails.

If this is not enough we might mention explicitly that Tails runs
independently from the original OS.