Fix panic mode on MAC spoofing failure, bis repetita.
In the past, we had "exit 1" in there. This problem was identified (#8571) and
fixed (commit 1b46b5b) in Tails 1.2.3, by replacing this "exit 1"
with "return 1".

Then, while working on another, minor problem (#8687), the "exit 1" was
reintroduced (commit 4ea050a) in Tails 1.3.2, presumably because we pasted
sample code from the ticket, that had been drafted *before* #8571 was fixed.
Sadly, nobody noticed in time.

I cannot easily reproduce the bug on Tails/Wheezy (because I lack hardware whose
MAC address fails to be spoofed there), but I could reproduce it with
Tails/Jessie ("thanks" to the fact we're currently shipping the buggy 'wl'
kernel module in there).

Will-Fix: #9531