Test suite: run ping as root.
For some reason, on Jessie, running ping as a regular users results in "ping:
icmp open socket: Operation not permitted", with exit code == 2. But as root, it
"works" and the firewall blocks the packets. This is rather an improvement than
a problem (stuff is blocked earlier, which is cheaper), so let's just deal with
it in the test suite only, by running ping as root: the main purpose here is to
test the firewall.

This change also affects the netcat command used to open TCP and UDP
connections, for code simplicity's sake. Here again, the goal is to test
the firewall.