Allow running the test suite without ntp installed.
There are other means to have an accurate host system clock, e.g.
systemd-timesyncd and tlsdate.

Still, keep installing ntp in the test suite setup documentation, so it results
in a working environment by default, and clarify in that same piece of
documentation that alternate means of synchronizing the clock work as well.

Refs: #9651