Override Torbirdy defaults to match Tails' expectations
We want stream isolation so that is set here. Additionally we want to
use the standard SKS keyserver pool like the command-line gnupg client
and Seahorse use in Tails.

Note that enigmail does *not* support using hkps.

Note also that we patch the source to due to the following as noted in
`components/torbirdy.js`:

// Default preference values for TorBirdy.
// These preferences values will be "enforced": even if the user decides to
// change the preferences listed below, they will be reset to the TorBirdy
// default when Thunderbird restarts. The reason we are doing this is because
// these preferences, if changed, can introduce leaks and therefore should be
// not changed by the user. Even if the user does change them, we reset them to
// the secure default when Thunderbird starts.