Dissimulate additional space introduced by ikiwiki (#9279)
The inline feature of ikiwiki is really not meant as a way of reusing
short pieces of text like this. Here is add a newline for each of them
which is then converted as a additional space by the browser. This
commit uses an ugly CSS trick to try to dissimulate this somehow.