Generate localized Wikipedia search engine plugins.
The new (ESR38) search bar only shows icons, which is problematic when
we want to include several locales of some search engine at the same
time, like Wikipedia (we want to include English in non-English
locales). Now we also generate localized Wikipedia search engine
plugin icons, which has an indicator of which language is used (by
language code) which should mitigate this.

Will-fix: #9955