apparmor-adjust-cupsd-profile.diff: adjust to parse fine on current Jessie.
Closes: #9963

More specifically:

a) avoid "conflicting x modifiers" for /usr/bin/hpijs, by modifying the
   /usr/bin/* rule to not match it;
b) add missing backends to the list of confined ones, and:
   asked how we can better maintain this list :
   https://lists.ubuntu.com/archives/apparmor/2015-August/008463.html
c) to avoid "conflicting x modifiers", replaced glob that matches
   all remaining backends by a hard-coded list of third-party ones
   we ship;
d) Added a note to the RM role duties to sanity check these two lists.

Also, add the attach_disconnected flag to the third_party local profile in
there, as was done in sid already, and is needed for it to work under systemd
(this patch already did that for /usr/sbin/cupsd).