Make --old-iso default to --iso if omitted.
Using the same ISO for the USB upgrade tests most often still does
what we want, e.g. test that the current version of Tails being tested
has a working Tails installer. Hence this seems like a reasonable
default. (Also note that an --old-iso may have changes that make it
impossible to run in the next versions automated test suite, and then
this is almost a requirement.)

Will-fix: #10147