Fix a corner case for the MAC spoofing panic mode.
If panic mode failed to disable the specific device that couldn't be
spoofed (by unloading the module) we disable networking. Previously we
only stopped NetworkManager. The problem is that NM isn't even started
at this time, but will specifically be started when we're done with
MAC spoofing.

Therefore, let's remove NetworkManager so it cannot possibly be
started.

Will-fix: #10160