Prevent unneccesssary bootstrap traffic when testing console failure modes
The change in this commit will simply disconnect the network right after
it tries to start reseeding. We block the router console port in this
scenario but I2P (upstream) will still run even if the router console is
not accessible for some reason.

Our NM hook starts I2P with `tails-i2p start` which will kill the
running I2P process if the router console is not accessible and that's
what this scenario tests.