Remove notes
Our table is only meant to provide useful hints. If the proposed keys are not
enough and people have to do more complicated actions than this, adding
small notes will probably not be enough anyway. We should have catch
all instructions for people to consult the user guide of their computer
instead.