Implement a more robust notification wait():er.
At least in the MAC spoofing .feature, it constantly happens that the
notification is lost even though notification-daemon is running. It
likely has to do with the notification being sent too early, before
the Tails Gnome session is ready for it.

Any way, it seems we need a more robust way for waiting for
notifications, given that they disappear by themselves, and I've also
observed this happen (especially when the host system is under heavy
load). So, the basic idea is to instead of the notification *pop-up*
look for the notification in the applet's list, when clicking on it.