Refactor "retry Tor" code.
There's some code simplification too, and the removal of the
@new_circuit_tries counter, which is all over the place and is a bit
hard to keep track of.

Also for the wget/whois test we do not try to detect if Tor was the
problem and only force a new circuit (and hence count it as a retry),
since that could lead to an infinite loop. With the new helper, we
always do MAX_NEW_TOR_CIRCUIT_RETRIES number of retries.