Add an improved debugging system via the pretty_debug formatter.
The idea is to replace each instance where we print debugging info to
STDERR if $config["DEBUG"] is set, with a call to debug_log(), which
will make the pretty_debug formatter print the output *if* it is
actiaved.

This approach is cleaner than what we do now, and will make it
possible to e.g. usw the pretty formatter for the terminal output and
at the same time logging pretty_debug to a file. Upon error, the full
log can then be consulted, without making the terminal output
messy. While debugging/developing the test suite, the pretty_debug
formatter can still be directed to the terminal, of course.

A notable problem is how to deal with the various subprocesses/threads
whose STDERR output we also are interested in when debugging only,
like Sikuli's debugging info. We'll have to find a way to capture such
output and hook it into debug_log() some how.