Make the -profile Tor Launcher workaround simpler.
... instead of the crazy workaround we had to use before. It turns out
you *can* use -app and -profile together, but only if -profile is
given last. It may be that it was fixed recently, because I'm pretty
sure it didn't work last time I wroked on this.

Also, in /usr/share/TorBrowser/Data/Browser (which is the "default"
profile directory relative to the Tor Launcher applocation.ini file)
the Caches directory must exist and be accessible for the tor-launcher
user even if -profile is used, so we just have to ensure it exists.

Will-fix: #7943