Stop installing system-config-printer.
It's been there since the first commit, mostly because back in the
years, it was the only decent CUPS GUI available; nowadays, GNOME's one
is good enough; we explicitly document that one should use the latter
anyway, and as shown by #8443, even us are sometimes confused by (or
reporting confusing information about) the fact that we ship 2 GUIs for
the same task.

Besides, on Jessie system-config-printer adds "an unclear and useless
Sundry category in Applications menu".

Note that typing in the sudoer password is still required to manage
printers, but that's not a Jessie regression and it's tracked by

Closes: #8505