Revert "Add a temporary checkpoint for the windows camouflage feature."
This reverts commit e13ee2563dd9063a4cae725cb72e39ec3fb82a52.

In the 'I enable Microsoft Windows camouflage' step we set @theme, but
it will only be set for a scenario where the checkpoint is created
(e.g. the first one run that uses it). To store such variables along
with the checkpoints we need #5847, so until then we cannot do
this. (Note that with the old system we could do this because we could
set it before the skip instruction in the step.)

Conflicts:
	features/step_definitions/snapshots.rb