Drop an assert whose usefulness seems hard to explain.
Previously, we were assuming that the Ruby bindings for virStorageVolDelete,
upon failure, convert the return code of this function into
a suitable exception.

This makes sense because I'm told that these bindings are (rightly, if one asks
me) built around this Ruby-ish principles.

Let's then assume the same for virStorageVolLookupByName.