Don't print which ISOs were testing.
We ought to know which ISOs we use from the command line. And we
certainly do not need to print them for every feature, like we did.

Besides, there's no good way for us to print this only once at the
start so that cucumbers formatters would pick it up. At best we could
print it right before the first "Scenario: ..." instance, but that
seems ugly.