Log useful information via Cucumber's formatters.
By writing to stderr such information is not logged to files when
using `--out`, but when writing to stdout (which `puts` does by
default) the human-readable formatters often do something smart (at
least 'pretty' and 'debug' do, which are the ones we care most about).

We only continue writing to stderr for the output of the options that
requires interaction, i.e. --retry-find and --pause-on-fail.

Will-fix: #10342