Rework how we measure pattern coverage.
The approaches we used before just seems incorrect in multiple ways:

* vm.overcommit_memory = 2 seems to make OOM killing happen way too
  early, around when 15% or RAM is still free, no matter the value of
  vm.overcommit_ratio and the various *reserved settings.

* Using the ratio of a ration for the {min,max}_coverage calculations
  were just weird and a sign of incremental bandaid-like fixes.

Now we do what we intended more explicitly, that is: measure the
pattern coverage in the free RAM at the time we start filling it. We
also have to reserve some RAM for the kernel (so it doesn't freeze
when approaching full RAM) and admin processes like the remote shell
(so it doesn't get unresponsive), and we exclude this from the free
RAM in the calculations for a much more accurate result than before.

Note that we may end up having a bit more than 100% pattern
coverage. While that may seem strange it may just indicate that some
unrelated RAM was freed after we noted the amount of free RAM right
before we started filling it.

Will-fix: #9705