XXX: Use a custom debootstrap script.
Setting this variable makes it possible to keep on using wheezy across
the whole live build process, except for the debootstrap step. This
modified copy of the wheezy script tweaks the environment before
live-build moves from the bootstrap step to the chroot step (package
installation etc.).

NOTE: Another way to implement this would be to insert some live-build
hooks between bootstrap and build steps, but it appears such hooks
aren't supported.

XXX: Having a whole copy made development easier, but on the long run it
would be better to apply a patch to the wheezy script to ensure further
maintenance.

Signed-off-by: Cyril Brulebois <cyril@debamax.com>