Fix for real the logical bug for real-only persistence.
The problem was not that live-persist must error out when read-only
persistence is activated rather than issue a warning. In the bug raised
by #10431, the problem was that it *already did* because it tried to
make the new persistence config anyway. So fixing #10431 by erroring out
just lead to the same situation.

In this case, we don't want to error out but warn out that the migration did not
happen, and just go on with the boot without trying to migrate anything.

Will-fix: #10431