Remove scenario.
In Jessie, org.gnome.gnome-screenshot.auto-save-directory is empty
since the default is sensible, ~/Pictures. To save the removed
scenario we'd have to set it explicitly, which seems stupid. Besides,
the 'A screenshot is taken when the PRINTSCREEN key is pressed'
scenario is already testing that GNOME Screenshot saves into this
directory, which is enough.