Turn htpdate.service into Type=oneshot.
I want to use time-sync.target (see systemd.special(7)), so that we can
order stuff after it.

But with Type=simple, we can't tell when htpdate is done, so we can't
specify that time-sync.target has not been reached until then. So let's
use Type=oneshot. But then, 20-time.sh would block until htpdate is
done, which is not what we want; this is solved by using --no-block when
restarting the service there.