Explicitly declare htpdate.service as being needed for time-sync.target.
This ensures that "services where correct time is essential should be
ordered after this unit" (as long as such services have themselves
declared the right dependencies). We don't ship any such services
currently, so this is rather pedantic a commit.

Note: in our case, we don't want the Tor service to wait for
time-sync.target, since we still use the Tor consensus for time
sync' purposes.