Allow step to be retried
If we don't monitor whether the key has been fetched with the gpg CLI
tool we can move to the next step which used to check whether the key is
in the keyring. By that time it's too late to retry fetching.

Instead let's check with `gpg` for the key. If the key is not in the
keyring within two minutes we'll fail the step, force a new tor circuit,
and try retrieving the key again.