MAC spoofing failure notification: remove the link to the documentation.
It was broken on Tails/Wheezy already (the link to the doc is not
visible, and one of the two possible links was broken anyway), so this
is not a regression brought by porting to Jessie. And on Jessie, due to
bug #7989 these links would not work as-is anyway. It's unclear how this
would be best solved; #10559 has been created to sum up the problem and
track future improvements.

Refs: #7989, #10559