Don't try to disable lvm2 initscripts anymore.
It was disabled in amnesia 0.3 because "it slows-down too much the boot
in certain circumstances". I can't confirm that with current hardware on
feature/jessie: the output of `systemd-analyze critical-chain' does not
include anything LVM-related, and in `systemd-analyze blame' everything
LVM-related takes 400ms or less on one of my systems, and 100ms or less
on another.

On Jessie, lvm2 initscripts have been split into numerous systemd units,
so it might help avoid unconditionally blocking boot.

Note that the way we did in this this chroot_local-hooks has no effect
on Jessie.