Work around issue with lost key presses in the browser.
When entering an address like"http(s)" we sometime end up with
"htp(s)", so either the second or third character is lost. My guess is
the second; once the first character is entered, the Tor Browser
immediately starts showing the list of suggestions, and it seems if
the second character is timed at some particular point when this list
is being shown, it's lost.

This is worked around by just pasting the whole address in one go, via
the clipboard.

Will-fix: #10467