Revert "Display date in the clock, in addition to time."
This reverts commit 82f5294b6730c3af29fe4a5d26ee7997a1c4a184.

I see no obvious reason to diverge from GNOME's defaults here.

In particular, as long as we are displaying time in UTC, which is
already misleading, the date can be different from what it is
in local time, which has even bigger potential for confusion.

And I also suspect that making the top-right menu narrower has chances
to improve things for #10576 and similar, at least in some cases.