Test suite: run ping as root.
On Jessie, setcap is used by default instead of setuid root for /bin/ping,
but aufs does not support file capabilities:

  $ /sbin/getcap /bin/ping
  Failed to get capabilities of file `/bin/ping' (Operation not supported)

  $ /sbin/getcap /lib/live/mount/rootfs/filesystem.squashfs/bin/ping
  /lib/live/mount/rootfs/filesystem.squashfs/bin/ping = cap_net_raw+ep

We could of course make /bin/ping setuid root back, just as it has
always been, but with our firewall it'll only allow pinging the LAN; for
now, I'm deciding that the limited usefulness is not worth the security
implications (even though we confine ping with AppArmor), and ping will
remain root only for now. We'll see how much sensible complains we get
during the 2.0 beta and RC phases.