Don't propose subscribing to the mailing list in final recommendations
At this point people are tired and not paying much attention anymore, so
these need to be the shortest possible.

Now that we have automatic upgrades, subscribing to amnesia-news is not
so important anymore, and people are shown with the news page when
opening Tor Browser anyway (which includes the same form about
subscribing to amnesia-news).