Remove superfluous part of test.
We still do the essential part i.e. test that Totem can open videos
over Tor. The removed part is essentially a test of the GUI, and what
a confusing and hard-to-discover GUI it is! I doubt users will ever
find out how to do it, so I fail to see why we should test it (and
waste time dealing with the robustness issue that plague this test).