Introduce a new XML-based iptables parser.
The old `iptables_parse()` way relies on regex parsing the output of
the `iptables` output, which of course is not very reliable. In Jessie
the `ip6tables` output has changed and differs (if there are no
options the "opt" column will be empty, and not have a "--"). The new
parser fixes this.

The goal will be to convert all `iptables_parse()` instances to the
new parser.

Will-fix: #9704