Lower waiting time for USB installation in the test suite.
So far we were waiting up to one hour, which is just the same as our
Jenkins inactivity timeout, so in practice when Tails Installer fails
and displays an error message, instead of reporting that the job failed
(which is the point of the exercise) we abort the job due to this
timeout (which communicates waaaay less clearly to me, with my Tails
Installer developer hat, that there's probably a bug I should fix, as
opposed to "here's a transient failure caused by instabilities of our
Jenkins setup"). My understanding is that this is what prevented the
Tails Installer bug described on #10717 from being identified earlier,
and I don't want to see that happen again.

IIRC this 1h timeout was relevant for running with nested virtualization
on slightly old (ThinkPad X200) hardware. I don't think anyone does that
anymore, and I would bet that on such hardware, lots of other bits of
our test suite will time out anyway.

Fix-committed: #10718