Bump timeout
After lots of testing I've found that I2P will generally fail within 80
seconds if it's going to fail. Recently I discovered that sometimes,
however, it takes a bit longer. The page is still trying to load and we
press ESC, the browser may be at a stage in which the URL we were trying
to load will not be in the address bar. Pressing reload in this case
will be for naught.

After a fair amount of testing I have not seen it take longer than 120
seconds. This means that we should always be at a state in which
pressing reload will actually be able to reload the page.