Make sure the kernel doesn't starve from memory during memory erasure.
We want to set it to the lowest possible value, to maximize the coverage
of memory erasure. But Documentation/sysctl/vm.txt for
vm.min_free_kbytes says: "if you set this to lower than 1024KB, your
system will become subtly broken, and prone to deadlock under high
loads". With 2048KB I've seen freezes, so let's not modify the kernel
default (8192KB on x86_64 currently).

Refs: #9707, #10487.