Use a shorter display for tails_ticket
The previous display was noisy in most situations and at least I used to
force the shorter #XXXX display in many places, like reports, etc.

I checked in many places that removing the leading 'ticket ' in the
display didn't break the formatting of previous shortcuts (including the
the doc and past reports and news).