Avoid saying "minimal Tails"
During our tests, people were not relating well "minimal" (here) with
"temporary" (or "intermediary") in the steps. They were not sure what
the end product was an whether it was safe to use.

This commit makes sense combined with 88a2e5d that clarifies the
properties of the temporary Tails once created.