Disable I2P's time syncing support
I2P is allowed full access to the Internet but any DNS lookups must be
done through Tor. I2P can block when the time synching is in progress
(upstream bug filed) and this can be triggered when the lookup is done
through a faulty Tor circuit.

Not only is this feature extraneous in Tails since we handle our own
time synchronization, disabling it should help improve robustness.