When (re-)setting up the network, let the rest of the system know that Tor is not working at the moment.
This matters e.g. if we have already bootstrapped: if we don't do that,
then tails-tor-has-bootstrapped.target is still active, so
tor-has-bootstrapped will return true as soon as tor@default.service is
active, that is as soon as it has open its ports... but it might not be
ready to build circuits yet.

Refs: #10732