Release process: drop duplicate section.
I've reintroduced it by mistake when resolving the last set of merge
conflicts. Note that we can't build from the tag the almost-final image
with this branch in, since we need to build that image before we even
know what we will put in the tagged snapshots of upstream APT
repositories, so before we create the tag.