Make it possible to eject CD-ROM from running VMs.
I looked at how virsh does it (the --eject option), and it removes the
<source> element (specifically *without* the DEVICE_MODIFY_FORCE, so
we drop it), claiming it is "the documented way" but I failed to find
any reference to this. Any way, it works, although Libvirt thinks it
doesn't, throwing in error that we carefully has to match and ignore.