Tor Browser download test: explicitly wait for the "Save as" button.
We were waiting for the dialog (BrowserDownloadDialog.png) to show up, and then
clicking on a button that is part of that dialog
(BrowserDownloadDialogSaveAsButton.png). In the past we've seen such
implementations be fragile, due to race conditions between when all the widgets
in a window are drawn vs. when Sikuli notices the first ones appear.

So, let's consider that "I get the browser download dialog" really means "I get
the browser download dialog including all its widgets that we are really
interested in".

refs: #9285