Seahorse test suite: update expected OpenPGP Applet icon to work even when it's truncated.
While running tests for test/10782-restart-seahorse-in-same-manner,
I've seen the "I start Seahorse via the Tails OpenPGP Applet" step
fail every time due to "FindFailed: can not find GpgAppletIconNormal.png
on the screen", apparently because the OpenPGP Applet icon was a tiny
bit truncated. With this small change to the image, it passes just fine.

refs: #10782