Simplify test.
Instead of waiting a long time, let's just check that networking truly
is disabled, in the same way we do for the MAC spoofing tests -- after
all the code for the "Diable networking" feature actually originates
from the MAC spoofing feature. This way we do not have to introduce a
static wait for 120 seconds, and we do not need the anti-test.