Test suite: bump a bit the timeout used when waiting for the remote shell.
We still see "Remote shell seems to be down" caused by a timeout,
and last time it happened the system was under heavy load.

I would like to discard any chance that we're simply not waiting long
enough. Given we wait up to 30 minutes for the system to boot, we can as
well wait for a few more dozens seconds for the remote shell to start.