Test suite: bump timeout for the process to disappear, when closing Evince.
I've seen it fail at least once in a way that suggests that waiting
a bit longer might have been enough:
https://jenkins.tails.boum.org/job/manual_test_Tails_ISO_testing/4/

So let's seize this opportunity to improve the test suite robustness
a bit. Worst case, if I was guessing wrong, we'll waste 10 seconds when
the tested functionality fails for real, which should be extremely rare.