Umount remaining tmp directories from old builds.
We've seens cases where builds on Jenkins left over some of the tmpfs
and aufs mounts when the job was removed while building, causing
failures of the next builds. This patch try to umount this kind of
mounts if there are, in hope to workaround this issue. It's made at the
very beginning of the build so that it fails fast and early in case the
umounting doesn't work.

Will-fix: #10772