Move requirements for Debian and Ubuntu version to the beginning of the scenario (#11017)
It's not nice to tell people they cannot finish a scenario after they started it.

We could even move this information in the /install/os but this would:

  - Require more design and implementation.
  - Might be harder to make as visible and information without polluting this
    page for everybody else.

Anyway, in a couple of months, Debian Wheezy will be unsupported officially and
Trusty will be superseded by a new LTS so we don't need a super fancy long-term
solution.