More robust and generic old build directory listing in the build script.
The previous one didn't catch the tmpfs build directories when built in
Ram. Using `tac` instead of `sort -r` also ensure that the listed
directory will be correctly reverse sorted, and so the umounting will
be done in the correct order if things like devpts or sysfs filesystems
are still mounted in the tmpfs.

Refs: #10772